+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-agent: a cleanup job to remove old Docker images (keeps two
+latest versions).
 - concord-server: an endpoint to download a single state file.